move amountAndPriceBox to parent class
